backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
langgraph
langchain-openai
langchain_groq
pydantic
python-multipart
